ML_function/app/requirements.txt
transformers
sentence-transformers
joblib
scipy
numpy
scikit-learn
streamlit
docx2txt
PyPDF2
python-docx
pdfplumber
nltk
torch